auth_service/requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.41.0
pydantic[email]==2.13.4
asyncpg==0.31.0
redis==7.4.0
aiokafka==0.14.0
boto3==1.42.58
motor==3.7.1
pymongo==4.15.4
elastic-apm[starlette]==6.25.0
elasticsearch[async]==8.19.3
SQLAlchemy==2.0.49
psycopg[binary]==3.2.13
alembic==1.17.2
argon2-cffi==25.1.0
PyJWT==2.10.1
httpx==0.28.1
python-multipart==0.0.20
pytest==9.0.2
pytest-asyncio==1.3.0
ruff==0.13.2
